backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
torch
numpy
asyncpg
pillow
python-multipart
aiohttp
python3-discogs-client
spotipy
python-dotenv
musicbrainzngs